+++ src/__init__.py
@@ -6,3 +6,4 @@ PARENT_FOLDER = str(Path(__file__).parent)
 DATA_FOLDER = os.path.join(ABSOLUTE_FOLDER, "data")
 TEST_FOLDER = os.path.join(DATA_FOLDER, "test")
 DEBUG_FOLDER = os.path.join(DATA_FOLDER, "debug")
+TEMPLATES_FOLDER = os.path.join(DATA_FOLDER, "templates")
